Update QM9 to allow for min heavy atoms
When performing graph training, we pass messages through bonds,
essentially. Not only are they overly trivial, but graphs in which
we have one node and no edges appear to present a specific edge
case problem for the graph based networks. Thus, there is now a flag
that allows the user to set a minimum number of heavy atoms. This
default is now 2.

Also add more tests.