Reformulate the batching method
The parallel model implementation in pytorch fails on the GPU for
an unknown reason. We suspect that it is because the GPU method
naively tries to break down a dgl.batch into sub-batches by
evenly distributing that across multiple GPU's but it doesn't know
that there are graphs of different sizes in the batch. So we end
up with situations where graphs are improperly split between
batches.

This commit hopefully fixes that problem by batching the graphs
as simple lists and performing the dgl.batch operation right before
executing forward prop.

Currently untested.